HCK-10091: Allow RE
