+++ api/re.js
@@ -1,3 +1,4 @@
 const { adaptJsonSchema } = require('../reverse_engineering/adaptJsonSchema');
+const { reFromFile } = require('../reverse_engineering/reFromFile');
 
-module.exports = { adaptJsonSchema };
+module.exports = { adaptJsonSchema, reFromFile };
